+++ src/yulplus.js
@@ -25,9 +25,9 @@ function id(x) { return x[0]; }
     ",": ",",
     ":": ":",
     MAX_UINTLiteral: /(?:MAX_UINT)/,
-    StrLiteral: /(?:str)"(?:\\["bfnrt\/\\]|\\u[a-fA-F0-9]{4}|[^"\\])*"/,
     SigLiteral: /(?:sig)"(?:\\["bfnrt\/\\]|\\u[a-fA-F0-9]{4}|[^"\\])*"/,
     TopicLiteral: /(?:topic)"(?:\\["bfnrt\/\\]|\\u[a-fA-F0-9]{4}|[^"\\])*"/,
+    ErrorLiteral: /(?:error)"(?:\\["bfnrt\/\\]|\\u[a-fA-F0-9]{4}|[^"\\])*"/,
     codeKeyword: /(?:code)(?:\s)/,
     objectKeyword: /(?:object)(?:\s)/,
     dataKeyword: /(?:data)(?:\s)/,
@@ -152,6 +152,11 @@ function id(x) { return x[0]; }
     return d;
   }
 
+  function functionDeclaration(d) {
+    d[2].type = 'FunctionIdentifier';
+    return d;
+  }
+
   function addValues(vals) {
     let cummulativeValue = utils.bigNumberify(0);
     let _vals = [0];
@@ -234,6 +239,12 @@ function require(arg) {
 
   // Include safe maths
   let identifierTree = {};
+
+  // idLiteral
+  function idLiteral(x) {
+    x[0].isLiteral = true;
+    return x[0];
+  }
 var grammar = {
     Lexer: lexer,
     ParserRules: [
@@ -370,6 +381,7 @@ var grammar = {
               throw new Error(`${_type} already declared with the same identifier "${v.name}" on line ${duplicateChecks[_type + v.name].line} and ${v.line}. All ${_type} must have the unique names. Scoping coming soon.`);
             }
           };
+          let replaceLiterals = {};
         
           let _map = mapDeep(d, v => {
             if (err) { throw new Error(err) }
@@ -386,7 +398,12 @@ var grammar = {
             // Set constants in context to used
             if (v.type === 'Constant') {
               v.type = 'UsedConstant';
-              v.block = currentBlock.id;
+        
+              if (v.__value.isLiteral) {
+                Object.keys(v.dataMap).map(identifier => {
+                    enums[identifier] = v.__value.value;
+                });
+              }
             }
         
             if (v.type === 'UsedConstant') {
@@ -422,6 +439,7 @@ var grammar = {
               && typeof enums[v.value] !== "undefined") {
         
               // Replace out enums
+              v.type = 'Literal';
               v.value = enums[v.value];
               v.text = enums[v.value];
             }
@@ -545,15 +563,16 @@ var grammar = {
           return { type: 'HexNumber', value: val, text: val };
         }
         },
-    {"name": "StrLiteral", "symbols": [(lexer.has("StrLiteral") ? {type: "StrLiteral"} : StrLiteral)], "postprocess": 
+    {"name": "ErrorLiteral", "symbols": [(lexer.has("ErrorLiteral") ? {type: "ErrorLiteral"} : ErrorLiteral)], "postprocess": 
         function(d) {
-          const abi = new utils.AbiCoder();
-          const str = abi.encode(['string'], [d[0].value.trim().slice(4).slice(0, -1)]); // remove str" and "
-          return {
-            type: 'HexNumber',
-            isString: true,
-            value: str,
-            text: str,
+          const message = d[0].value.trim().slice(6).slice(0, -1);
+          const bytes4ErrorHash = utils.keccak256(utils.toUtf8Bytes(message)).slice(0, 10); // remove error" and "
+          return { type: 'HexNumber',
+            isError: true,
+            hash: bytes4ErrorHash,
+            message,
+            value: bytes4ErrorHash,
+            text: bytes4ErrorHash,
           };
         }
         },
@@ -644,11 +663,11 @@ var grammar = {
     {"name": "NumericLiteral", "symbols": [(lexer.has("NumberLiteral") ? {type: "NumberLiteral"} : NumberLiteral)], "postprocess": id},
     {"name": "NumericLiteral", "symbols": [(lexer.has("HexNumber") ? {type: "HexNumber"} : HexNumber)], "postprocess": id},
     {"name": "NumericLiteral", "symbols": ["SigLiteral"], "postprocess": id},
+    {"name": "NumericLiteral", "symbols": ["ErrorLiteral"], "postprocess": id},
     {"name": "NumericLiteral", "symbols": ["TopicLiteral"], "postprocess": id},
-    {"name": "NumericLiteral", "symbols": ["StrLiteral"], "postprocess": id},
-    {"name": "Literal", "symbols": [(lexer.has("StringLiteral") ? {type: "StringLiteral"} : StringLiteral)], "postprocess": id},
-    {"name": "Literal", "symbols": ["NumericLiteral"], "postprocess": id},
-    {"name": "Literal", "symbols": ["MAX_UINT"], "postprocess": id},
+    {"name": "Literal", "symbols": [(lexer.has("StringLiteral") ? {type: "StringLiteral"} : StringLiteral)], "postprocess": idLiteral},
+    {"name": "Literal", "symbols": ["NumericLiteral"], "postprocess": idLiteral},
+    {"name": "Literal", "symbols": ["MAX_UINT"], "postprocess": idLiteral},
     {"name": "Expression", "symbols": ["Literal"], "postprocess": id},
     {"name": "Expression", "symbols": [(lexer.has("Identifier") ? {type: "Identifier"} : Identifier)], "postprocess": id},
     {"name": "Expression", "symbols": ["FunctionCall"], "postprocess": id},
@@ -707,6 +726,8 @@ var grammar = {
             const properties = _filter(d[6], 'MemoryStructIdentifier');
             let methodList = properties.map(v => name + '.' + v.name);
         
+            // console.log('mstruct', properties);
+        
             // check for array length specifiers
             for (var p = 0; p < properties.length; p++) {
               const prop = properties[p];
@@ -734,6 +755,11 @@ var grammar = {
           res := mslice(add(${name + '.' + v.name}.position(pos),
             mul(i, ${v.value.value})), ${v.value.value})
         }
+        
+        function ${name + '.' + v.name}.slice(pos) -> res {
+          res := mslice(${name + '.' + v.name}.position(pos),
+            ${name + '.' + v.name}.length(pos))
+        }
         `
         : `
         function ${name + '.' + v.name}(pos) -> res {
@@ -741,8 +767,23 @@ var grammar = {
         }
         `,
                 required: [
-                  name + '.' + v.name + '.position',
-                ],
+                  name + '.' + v.name + '.position'
+                ].concat(v.value.type === 'ArraySpecifier'
+                  ? [name + '.' + v.name + '.length']
+                  : []),
+              },
+              [name + '.' + v.name + '.slice']: {
+                method: v.value.type === 'ArraySpecifier' ?
+        `
+        function ${name + '.' + v.name}.slice(pos) -> res {
+          res := mslice(${name + '.' + v.name}.position(pos), ${name + '.' + v.name}.length(pos))
+        }
+        ` : ``,
+                required: [
+                  name + '.' + v.name + '.position'
+                ].concat(v.value.type === 'ArraySpecifier'
+                  ? [name + '.' + v.name + '.length']
+                  : []),
               },
               [name + '.' + v.name + '.keccak256']: {
                 method: `
@@ -859,6 +900,13 @@ var grammar = {
           }), {});
           d.__constant = true;
         
+          // if its a literal, we move for a global replacement block >
+          if (d[0].__value.isLiteral) {
+            d[0].value = '';
+            d[0].text = '';
+            return d[0];
+          }
+        
           return d;
         }
         },
@@ -870,11 +918,11 @@ var grammar = {
           return d;
         }
         },
-    {"name": "FunctionDefinition", "symbols": [{"literal":"function"}, "_", (lexer.has("Identifier") ? {type: "Identifier"} : Identifier), "_", {"literal":"("}, "_", "IdentifierList", "_", {"literal":")"}, "_", {"literal":"->"}, "_", "IdentifierList", "_", "Block"]},
-    {"name": "FunctionDefinition", "symbols": [{"literal":"function"}, "_", (lexer.has("Identifier") ? {type: "Identifier"} : Identifier), "_", {"literal":"("}, "_", {"literal":")"}, "_", {"literal":"->"}, "_", "IdentifierList", "_", "Block"]},
-    {"name": "FunctionDefinition", "symbols": [{"literal":"function"}, "_", (lexer.has("Identifier") ? {type: "Identifier"} : Identifier), "_", {"literal":"("}, "_", {"literal":")"}, "_", {"literal":"->"}, "_", {"literal":"("}, "_", {"literal":")"}, "_", "Block"]},
-    {"name": "FunctionDefinition", "symbols": [{"literal":"function"}, "_", (lexer.has("Identifier") ? {type: "Identifier"} : Identifier), "_", {"literal":"("}, "_", "IdentifierList", "_", {"literal":")"}, "_", "Block"]},
-    {"name": "FunctionDefinition", "symbols": [{"literal":"function"}, "_", (lexer.has("Identifier") ? {type: "Identifier"} : Identifier), "_", {"literal":"("}, "_", {"literal":")"}, "_", "Block"]},
+    {"name": "FunctionDefinition", "symbols": [{"literal":"function"}, "_", (lexer.has("Identifier") ? {type: "Identifier"} : Identifier), "_", {"literal":"("}, "_", "IdentifierList", "_", {"literal":")"}, "_", {"literal":"->"}, "_", "IdentifierList", "_", "Block"], "postprocess": functionDeclaration},
+    {"name": "FunctionDefinition", "symbols": [{"literal":"function"}, "_", (lexer.has("Identifier") ? {type: "Identifier"} : Identifier), "_", {"literal":"("}, "_", {"literal":")"}, "_", {"literal":"->"}, "_", "IdentifierList", "_", "Block"], "postprocess": functionDeclaration},
+    {"name": "FunctionDefinition", "symbols": [{"literal":"function"}, "_", (lexer.has("Identifier") ? {type: "Identifier"} : Identifier), "_", {"literal":"("}, "_", {"literal":")"}, "_", {"literal":"->"}, "_", {"literal":"("}, "_", {"literal":")"}, "_", "Block"], "postprocess": functionDeclaration},
+    {"name": "FunctionDefinition", "symbols": [{"literal":"function"}, "_", (lexer.has("Identifier") ? {type: "Identifier"} : Identifier), "_", {"literal":"("}, "_", "IdentifierList", "_", {"literal":")"}, "_", "Block"], "postprocess": functionDeclaration},
+    {"name": "FunctionDefinition", "symbols": [{"literal":"function"}, "_", (lexer.has("Identifier") ? {type: "Identifier"} : Identifier), "_", {"literal":"("}, "_", {"literal":")"}, "_", "Block"], "postprocess": functionDeclaration},
     {"name": "Empty", "symbols": [(lexer.has("space") ? {type: "space"} : space)]},
     {"name": "Empty", "symbols": [(lexer.has("multiComment") ? {type: "multiComment"} : multiComment)]},
     {"name": "Empty", "symbols": [(lexer.has("singleLineComment") ? {type: "singleLineComment"} : singleLineComment)]},